# notion/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Notion Tool API",
    "description": "This API provides endpoints for interacting with Notion, including loading pages or databases, searching for content, and appending data to pages or blocks.",
    "version": "0.1.0"
  },
  "paths": {
    "/read-page": {
      "get": {
        "summary": "Read notion page",
        "description": "Read notion page with given page_id",
        "operationId": "load_data_read_page_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "page_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "The Notion page ID to read.",
              "title": "Page Id"
            },
            "description": "The Notion page ID to read."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query-database": {
      "get": {
        "summary": "Query notion database",
        "description": "Queries a notion database with given database_id",
        "operationId": "load_data_query_database_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "database_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "The Notion database ID to load.",
              "title": "Database Id"
            },
            "description": "The Notion database ID to load."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/load": {
      "get": {
        "summary": "Load Notion pages or databases",
        "description": "Loads a list of pages or databases given their IDs or a single database ID.",
        "operationId": "load_data_load_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "page_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "List of Notion page IDs to load.",
              "default": [],
              "title": "Page Ids"
            },
            "description": "List of Notion page IDs to load."
          },
          {
            "name": "database_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The Notion database ID to load.",
              "title": "Database Id"
            },
            "description": "The Notion database ID to load."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Search for Notion pages or databases",
        "description": "Searches for matching pages or databases based on the query.",
        "operationId": "search_data_search_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Query string for searching Notion pages or databases.",
              "title": "Query"
            },
            "description": "Query string for searching Notion pages or databases."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/append": {
      "post": {
        "summary": "Append content to a Notion page or database",
        "description": "Appends text content to the end of the specified Notion page or block.",
        "operationId": "append_data_append_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AppendDataRequest": {
        "properties": {
          "block_id": {
            "type": "string",
            "title": "Block Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "text_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Type"
          }
        },
        "type": "object",
        "title": "AppendDataRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Api-Key"
      }
    }
  }
}